+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Unneeded line breaks when using `-o` in the `search` command. (#1425) (@fukusuket)
 - Sigma correlation rules required the `group-by` field but now it is optional. (#1442) (@fukusuket)
+- Hayabusa will give an error message if the rules referenced by a correlation rule are not found. (#1444) (@fukusuket)
 
 ## 2.17.0 [2024/08/23] "HITCON Community Release"